onboarding.us.put_financial_market_experience_time/requirements.txt
--extra-index-url https://backend:dc222fa6-452f-4253-a2fd-a52e7ba2e7e9@nexus.sigame.com.br/repository/pypi/simple
heimdall-client==2.3.1
persephone-client==2.2.0
etria-logger==0.1.10
python-decouple==3.6
mnemosine==0.1.2

motor==3.0.0
flask==2.1.2
iara-client==0.3.2
cx-Oracle==8.3.0
httpx==0.23.0

uvloop==0.17.0
uvicorn==0.20.0
asgiref==3.5.2
hypercorn==0.13.2